requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
Flask>=2.3.0
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
openai>=1.0.0
anthropic>=0.3.0
boto3>=1.26.0